+++ parser_library/src/checking/instr_operand.h
@@ -78,16 +78,25 @@ enum class machine_operand_type : uint8_t
     RELOC_IMM,
 };
 
+enum class even_odd_register : uint8_t
+{
+    NONE,
+    ODD,
+    EVEN,
+};
+
 // Describes a component of machine operand format. Specifies allowed values.
 struct parameter
 {
-    bool is_signed;
-    uint8_t size;
-    machine_operand_type type;
-
-    constexpr bool is_empty() const { return (!is_signed && type == machine_operand_type::NONE && size == 0); }
+    bool is_signed : 1;
+    uint8_t size : 7;
+    machine_operand_type type : 4;
+    even_odd_register evenodd : 2 = even_odd_register::NONE;
+    uint8_t min_register : 2 = 0;
 
     bool operator==(const parameter&) const = default;
+
+    constexpr bool is_empty() const { return *this == parameter {}; }
 
     std::string to_string() const;
 };